% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{belgium}
\alias{belgium}
\title{Serology and stratum proportion data from a
nationwide seroprevalenece study in Belgium}
\format{
\enumerate{
\item \strong{serology} A data frame with 22545 rows and 5 columns
\describe{
\item{age_cat}{age category of [0,10), [10, 20), etc.}
\item{sex}{female or male}
\item{province}{province}
\item{igg_cat}{positive or negative IgG result}
\item{collection_round}{Collection round between 1 and 7}
}

\item \strong{strataprops} A data frame with 220 rows and 5 columns
\describe{
\item{province}{province}
\item{age_cat}{age category of [0,10), [10, 20), etc.}
\item{sex}{female or male}
\item{count}{Number of persons in the stratum}
\item{stratum_prop}{Proportion of the population the stratum comprises}
}
}
}
\source{
\enumerate{
\item \strong{serology}
\describe{Herzog, S.A., De Bie, J., Abrams, S., et al., 2022.
Seroprevalence of IgG antibodies against SARS-CoV-2 - a serial prospective
cross-sectional nationwide study of residual samples, Belgium, March to October 2020.
\emph{Eurosurveillance}, 27(9), p.2100419.
\url{https://doi.org/10.2807/1560-7917.ES.2022.27.9.2100419},
\url{http://doi.org/10.5281/zenodo.4664403}
}
\item \strong{strataprops}
\describe{Federal Planning Bureau (2021). Population projections 2020-2070. Dataset.
\url{https://www.plan.be/databases/data-35-en-population_projections_2020_2070}
}
}
}
\usage{
belgium
}
\description{
This dataset is a list with two dataframes,
serology and strataprops. Both dataframes contain the
variables for standardization: age category (age_cat),
sex, province
}
\keyword{datasets}
