% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rg.R
\name{rg}
\alias{rg}
\title{Compute Rogan-Gladen estimate}
\usage{
rg(rho_hat, sigma_e_hat, sigma_p_hat, n_1, n_2, n_3, variance = TRUE)
}
\arguments{
\item{rho_hat}{Raw/apparent prevalence (positives divided by total samples)}

\item{sigma_e_hat}{Estimated sensitivity}

\item{sigma_p_hat}{Estimated specificity}

\item{n_1}{Sample size for sensitivity validation dataset}

\item{n_2}{Sample size for specificity validation dataset}

\item{n_3}{Sample size for main study from population}

\item{variance}{Should variance be estimated? (default TRUE)}
}
\value{
A list of estimated prevalence and variance; optionally,
if variance is FALSE, the truncated prevalence estimate
\itemize{
\item hat_pi_rg - Estimated Rogan-Gladen prevalence
\item hat_var_pi_rg - Estimate of the variance of the Rogan-Gladen estimator
}
}
\description{
The estimator corresponds to the point and variance estimators
from Rogan and Gladen (1978)
}
\examples{
rg(0.01, .99, .98, 30, 50, 1000, TRUE)

}
