% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srg.R
\name{srg}
\alias{srg}
\title{Compute standardized Rogan-Gladen estimate}
\usage{
srg(
  sample,
  stratum_props,
  sigma_e_hat,
  sigma_p_hat,
  n_1,
  n_2,
  n_3,
  vars_std,
  variance = TRUE
)
}
\arguments{
\item{sample}{A dataframe with data from the sample, including the stratum_props (\eqn{\gamma_j} values) which will be used in standardization}

\item{stratum_props}{The stratum_props dataframe}

\item{sigma_e_hat}{Estimated sensitivity}

\item{sigma_p_hat}{Estimated specificity}

\item{n_1}{Sample size for sensitivity validation dataset}

\item{n_2}{Sample size for specificity validation dataset}

\item{n_3}{Sample size for main study from population}

\item{vars_std}{The names of variables to standardize over, corresponding to columns in sample}

\item{variance}{Should variance be estimated?}
}
\value{
A list of estimated prevalence and variance; optionally,
if variance is FALSE, the truncated prevalence estimate
\itemize{
\item hat_pi_srg - Estimated standardized Rogan-Gladen prevalence
\item hat_var_pi_srg - Estimate of the variance of the standardized Rogan-Gladen estimator
\item num_strata - Number of strata in the sample used for standardization
}
}
\description{
Computes nonparametric standardized Rogan-Gladen prevalence estimator.
Corresponds to \eqn{\hat{\pi_{SRG}}} and \eqn{\hat V_{\pi, SRG}} in the manuscript.
}
\examples{

 serology <- belgium$serology
 belg_cr <- serology[serology$collection_round == 1,]
 belg_cr$x <- ifelse(belg_cr$igg_cat=="positive", 1, 0)
 belg_cr_srg <- as.data.frame(belg_cr)

 srg(belg_cr_srg, belgium$strataprops, 154/181, 322/326,
      181, 326, 3000, vars_std = c("age_cat", "sex"), variance = TRUE)

}
